final_proj/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
docker==5.0.3
docker-pycreds==0.4.0
Faker==10.0.0
Flask==2.0.2
greenlet==1.1.2
h11==0.12.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.28
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
wsproto==1.0.0
